+++ next.config.ts
@@ -1,7 +1,12 @@
-import type { NextConfig } from "next";
-
-const nextConfig: NextConfig = {
-  /* config options here */
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  reactStrictMode: true,
+  swcMinify: true,
+  output: 'standalone', // needecd for docker server.js production setup
+  experimental: {
+    // If you're using App Router
+    serverActions: true,
+  },
 };
 
 export default nextConfig;