+++ src/components/animations/index.ts
@@ -1,0 +1,2 @@
+export * from './page-transition';
+export * from './fadeIn';